+++ project/app/(dashboard)/team/page.tsx
@@ -1,5 +1,5 @@
-import { AddTeamMemberButton } from "@/components/add-member-button";
 import { DashboardContent } from "@/components/layouts/dashboard/dashboard-content";
+import { AddTeamMemberButton } from "@/components/ui/buttons/add-member-button";
 import TeamCard from "@/components/ui/team-card";
 
 export default function TeamPage() {
