+++ project/components/sections/projects/projects-list.tsx
@@ -1,6 +1,6 @@
 "use server";
 
-import { CreateProjectButton } from "@/components/create-project-button";
+import { CreateProjectButton } from "@/components/ui/buttons/create-project-button";
 import ProjectCard from "@/components/ui/project-card";
 import { getProjects } from "@/lib/actions/projects";
